lab_2/neural_model/model/requirements.txt
torch
torchvision
timm
pillow
numpy
tqdm